+++ lanovenko.ivan/T2/data_struct.hpp
@@ -15,6 +15,24 @@ namespace lanovenko
     std::string key3;
   };
 
+  struct CharIO
+  {
+    char& ref;
+  };
+
+  struct RationalIO
+  {
+    std::pair< long long, unsigned long long >& ref;
+  };
+
+  struct StringIO
+  {
+    std::string& ref;
+  };
+
+  std::istream& operator>>(std::istream& in, CharIO&& dest);
+  std::istream& operator>>(std::istream& in, RationalIO&& dest);
+  std::istream& operator>>(std::istream& in, StringIO&& dest);
   std::istream& operator>>(std::istream& in, DataStruct& dest);
   std::ostream& operator<<(std::ostream& out, const DataStruct& dest);
   bool compare(const DataStruct& lhs, const DataStruct& rhs);
